ocr_backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
torch>=2.0.0
transformers>=4.40.0
Pillow>=10.0.0
ollama>=0.1.0
